Move I2C types to a crate; remove 'i2c' feature
