+++ drv/i2c-types/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "drv-i2c-types"
+description = "data types for the client I2C API"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+num-derive.workspace = true
+num-traits.workspace = true
+hubpack.workspace = true
+serde.workspace = true
+enum-kinds.workspace = true
+
+derive-idol-err.path = "../../lib/derive-idol-err"
